Refactor out flat curve
