+++ token-swap/program/src/curve/mod.rs
@@ -1,2 +1,4 @@
 //! ...
+
 pub mod shared;
+pub mod flat;
